autoinstall_snippets/pre_install_network_config: fixup undefined gateway
"gateway" is not a network interface field,

Rendering the sample.ks autoinstall template fails due to the variable
"gateway" being undefined. This happens in the pre_install_network_config snippet,
which reads the gateway from the interface's configuration dict
whereas this seems to be a "global" variable,
as used in the post_install_network_config snippet.

-> do not get "gateway" from idata, assume it's a global var